refactor columns implementation
